More efficient handling of Rods + major remodeling of the library
1) simulation.py - holds the Simulation class which is the main
interface for the user to the library and its primary functions.
Now Rod objects are created at start and kept in memory all the time as
the "state" of the simulation - allows easier random picking but also
possible further enhancements (removing, adding, etc.)

2) model.py - holds a Model class which is basically just a placeholder
for all the information that is given by the config file (which
basically contains the model for the simulation)

3) rod.py - now holds the Rod class, which is optimised by having a
reference to the Simulation object and uses its information about the
simulation instead of keeping its own (e.g. list of all particle types)